% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrGraphCloneAnalysis.R
\name{getClonesFromTcrGraph}
\alias{getClonesFromTcrGraph}
\title{Get clone information}
\usage{
getClonesFromTcrGraph(
  tcrData,
  maxA = 2,
  maxB = 2,
  maxD = -1,
  maxG = -1,
  maxALib = maxA,
  maxBLib = maxB,
  maxDLib = maxD,
  maxGLib = maxG,
  maxAGraph = maxA,
  maxBGraph = maxB,
  maxDGraph = maxD,
  maxGGraph = maxG,
  format = "compressed",
  link = "full_nt_sequence"
)
}
\arguments{
\item{tcrData}{Either a tcrGraph object or a data frame containing columns named libid, v_gene, j_gene, and the column 
indicated by the \code{link} argument.}

\item{maxA}{the maximum number of alpha chains in a clone. A value of -1 will cause alphas to be removed from consideration. Default: 2.}

\item{maxB}{the maximum number of beta chains in a clone. A value of -1 will cause betas to be removed from consideration. Default: 2.}

\item{maxD}{the maximum number of delta chains in a clone. A value of -1 will cause deltas to be removed from consideration. Default: -1.}

\item{maxG}{the maximum number of gamma chains in a clone. A value of -1 will cause gammas to be removed from consideration. Default: -1.}

\item{maxALib}{the maximum number of alpha chains in a library A value of -1 will cause alphas to be removed from consideration. Default: maxA.}

\item{maxBLib}{the maximum number of beta chains in a library A value of -1 will cause betas to be removed from consideration. Default: maxB.}

\item{maxDLib}{the maximum number of delta chains in a library A value of -1 will cause deltas to be removed from consideration. Default: maxD.}

\item{maxGLib}{the maximum number of gamma chains in a library A value of -1 will cause gammas to be removed from consideration. Default: maxG.}

\item{maxAGraph}{the maximum number of alpha chains in a subgraph A value of -1 will cause alphas to be removed from consideration. Default: maxA.}

\item{maxBGraph}{the maximum number of beta chains in a subgraph A value of -1 will cause betas to be removed from consideration. Default: maxB.}

\item{maxDGraph}{the maximum number of delta chains in a subgraph A value of -1 will cause deltas to be removed from consideration. Default: maxD.}

\item{maxGGraph}{the maximum number of gamma chains in a subgraph A value of -1 will cause gammas to be removed from consideration. Default: maxG.}

\item{format}{a string, one of "compressed", "full", or "tcrGraph". In "compressed" format, the returned data frame have one row for each 
clone, with all of the libs where that clone was found listed together in one column. In "full" format, each lib will be on a separate row, 
with a column for clone ID. In "tcrGraph" format, a tcrGraph object is returned. Default is "compressed".}

\item{link}{a string indicating which column to use to match chains. Default is "full_nt_sequence", but only used if \code{tcrData} is
a data frame object.}
}
\value{
A data frame of information about the clones detected in the TCR graph
}
\description{
Given TCR data, identify clones and count the number of occurrences for each clone. A clone is defined as a maximal subgraph
containing at most the number of alpha, beta, delta, and gamma chains set by `maxA`, `maxB`, `maxD`, and `maxG`. The arguments
`max[A,B,D,G]Lib` and `max[A,B,D,G]Graph` allow for more refined control over the steps at which chains are filtered. 
`Max[A,B,D,G]Lib` arguments cause libraries with more than the allowed number of chains to be removed before clones are counted.
`Max[A,B,D,G]Graph` arguments cause the nodes with the fewest number of counts to be removed from each maximal subgraph until 
the maximum number of allowed chains remain. This subgraph is then counted as a clone.
}
\examples{
\dontrun{
tcrData <- data.frame(
   libid = c("lib000", "lib000"), 
   v_gene = c("TRAV1-1", "TRBV1"), 
   j_gene = c("TRAJ11", "TRBJ1-1"), 
   fake_nt_sequence = c("CAT", "GAT")
)
myGraph <- makeTcrGraph(tcrData, link = "fake_nt_sequence")
getClonesFromTcrGraph(myGraph)
# A tibble: 1 x 6
#  cloneId cloneCounts libs   vGenes         jGenes          fake_nt_sequence
#  <chr>         <int> <chr>  <chr>          <chr>           <chr>           
#  clone1            1 lib000 TRAV1-1, TRBV1 TRAJ11, TRBJ1-1 CAT, GAT  
}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
