% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrGraphUtils.R
\name{makeTcrGraph}
\alias{makeTcrGraph}
\title{Make a tcrGraph object}
\usage{
makeTcrGraph(tcrDf, link = "full_nt_sequence")
}
\arguments{
\item{tcrDf}{a data frame that contains columns libid, v_gene, j_gene, and the column indicated by the \code{link} argument}

\item{link}{a string or vector of strings indicating which column(s) to use to match chains. Default is "full_nt_sequence"}
}
\value{
an S3 object of class \code{tcrGraph}
}
\description{
Accepts a dataframe containing TCR information, and generates a tcrGraph object
}
\examples{
\dontrun{
tcrData <- data.frame(
   libid = c("lib000", "lib000"), 
   v_gene = c("TRAV1-1", "TRBV1"), 
   j_gene = c("TRAJ11", "TRBJ1-1"), 
   fake_nt_sequence = c("CAT", "GAT")
)
makeTcrGraph(tcrData, link = "fake_nt_sequence")
# An object of class 'tcrGraph'
# Nodes defined by fake_nt_sequence
# Number of nodes: 2
# Number of edges: 2

makeTcrGraph(tcrData, link = c("v_gene", "fake_nt_sequence"))
# An object of class 'tcrGraph'
# Nodes defined by v_gene_fake_nt_sequence
# Number of nodes: 2
# Number of edges: 2
}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
