% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrGraphFigures.R
\name{makeVisNetwork}
\alias{makeVisNetwork}
\title{Generates an interactive visNetwork plot}
\usage{
makeVisNetwork(tcrGraph)
}
\arguments{
\item{tcrGraph}{the tcrGraph object to generate a plot from}
}
\value{
an interactive visNetwork plot
}
\description{
Generates a visNetwork plot with reasonable defaults for
data exploration. Investigating a TCR network in this way will
probably be more useful than using the base plotting function
for networks with mrore than a few nodes. Hovering over each
node will provide additional information about the chain.
}
\examples{
\dontrun{
tcrData <- data.frame(
   libid = c("lib000", "lib000"), 
   v_gene = c("TRAV1-1", "TRBV1"), 
   j_gene = c("TRAJ11", "TRBJ1-1"), 
   fake_nt_sequence = c("CAT", "GAT")
)
myGraph <- makeTcrGraph(tcrData, link = "fake_nt_sequence")
makeVisNetwork(myGraph)     
}
}
\seealso{
\code{\link[visNetwork]{visNetwork}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
