% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrGraphS3Methods.R
\name{plot.tcrGraph}
\alias{plot.tcrGraph}
\title{Plot method for tcrGraph objects}
\usage{
\method{plot}{tcrGraph}(x, nodeScalingFactor = 1, chainPalette = rainbow(4), ...)
}
\arguments{
\item{x}{object of class 'tcrGraph' to be printed}

\item{nodeScalingFactor}{a number to multiply the number of node occurrences by to generate the node size in the plot. Default 1.}

\item{chainPalette}{a list of 4 colors to be assigned to chains of type TRA, TRB, TRD, and TRG. Any names will be overridden. 
Default rainbow(4)}

\item{...}{additional arguments to be passed to igraph.plot(...)}
}
\description{
Creates simple plot of TCR graph structure
}
\examples{
\dontrun{
tcrData <- data.frame(
   libid = c("lib000", "lib000"), 
   v_gene = c("TRAV1-1", "TRBV1"), 
   j_gene = c("TRAJ11", "TRBJ1-1"), 
   fake_nt_sequence = c("CAT", "GAT")
)
myGraph <- makeTcrGraph(tcrData, link = "fake_nt_sequence")
plot(myGraph, nodeScalingFactor = 30)
}

}
\seealso{
\code{\link[igraph]{igraph.plotting}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
