% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrGraphS3Methods.R
\name{tcrGraph}
\alias{tcrGraph}
\alias{is.tcrGraph}
\alias{as.data.frame}
\title{tcrGraph S3 class}
\usage{
tcrGraph(data, nodes, edges, link)

## S3 methods for class 'tcrGraph'
is.tcrGraph(x)
}
\arguments{
\item{data}{data frame containing TCR data used to construct the nodes and edges. This must include columns named \code{libid}, 
\code{id}, \code{chainType}, and the column indicated by the \code{link} argument.}

\item{nodes}{data frame containing columns \code{id} and \code{group}, where group is one of "TRA", "TRB", "TRD", or "TRG"}

\item{edges}{data frame containing columns \code{from} and \code{to}, where both from and to values appear as node ids}

\item{link}{string indicating the column name in \code{data} used to define nodes (ie: link between libraries)}

\item{x}{object to be tested for class tcrGraph or coerced into a data frame}
}
\description{
Functions to create and check for tcrGraph objects.
}
\details{
Note that it is usually not advised to manually construct a tcrGraph object. To generate a tcrGraph from a data frame of TCR data, this
package provides the \code{\link{makeTcrGraph}} function, which should be used in most cases.

Note also that when a field named "title" is present in tcrGraph nodes and edges data, this field is assumed  
to contain summary data as HTML that can be rendered during a mouseover event in a browser. This is useful 
for some interactive plotting packages such as \code{\link[visNetwork]{visNetwork}}, but it means you should
avoid using 'title' as a field/column name in your TCR data frame.
}
\seealso{
\code{\link{plot.tcrGraph}}, \code{\link{print.tcrGraph}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
