% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrGraphUtils.R
\name{tcrGraphDegree}
\alias{tcrGraphDegree}
\title{Calculate the degree for each node of a tcrGraph obect}
\usage{
tcrGraphDegree(tcrGraph)
}
\arguments{
\item{tcrGraph}{the tcrGraph object to calculate the degree of}
}
\value{
a data frame containing the degree for each node id
}
\description{
The degree of a node is the number of connections it has to other nodes.
}
\examples{
\dontrun{
tcrData <- data.frame(
   libid = c("lib000", "lib000"), 
   v_gene = c("TRAV1-1", "TRBV1"), 
   j_gene = c("TRAJ11", "TRBJ1-1"), 
   fake_nt_sequence = c("CAT", "GAT")
)
myGraph <- makeTcrGraph(tcrData, link = "fake_nt_sequence")
tcrGraphDegree(myGraph) 
       degree     id
node_1      1 node_1      
node_2      1 node_2        
}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
