% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrGraphUtils.R
\name{tcrGraphToIgraph}
\alias{tcrGraphToIgraph}
\title{Convert a tcrGraph obect to an igraph object}
\usage{
tcrGraphToIgraph(tcrGraph)
}
\arguments{
\item{tcrGraph}{the tcrGraph object to convert}
}
\value{
an igraph graph object
}
\description{
Converting a tcrGraph to an igraph object enables the use of utility and 
statistics functions from the \code{link[igraph]{igraph}} package.
}
\seealso{
\code{link[igraph]{igraph}}
}
\author{
Mario G Rosasco, \email{mrosasco@benaroyaresearch.org}
}
